remove old Bu2019lm pkl, docs update
Removes the old Bu2019lm.pkl file in svdmodels and updates the documentation to include links to the oversized pickle files required to use the Bu2019lm model. Also added a note about creating a symbolic link between svdmodels folders and reorganized some of the referenced installation issues for clarity and changed some of bare urls to be hyperlinks.