lightcurve comparison implemented
all functions exist now such that it should be possible to test the models and compare if command line and imported methods for generating lightcurves can cause issues. Currently, only have nugent-hyper (supernova) as a test, recommend adding other common models to the lists on lines 266 and 267. May be better to have it as a passable argument and a defined dictionary containing all model names and associated functions.