Update corner_plot.py
refactored the `plotting_parameter` function. axis labels and parameters which needs to be plotted are calculated based on the intersection of prior and 1st posterior file (see note below). added `verbose` flag to print relevant information to help debugging. 

Note: the code assumes that the parameters in rest of the posterior files are the same as the first posterior file. and the prior file and posterior files have the same parameters which can be plotted