relocate try except
moves the try except for loading lightcurve files into the loadEvent function in io.py. This is so calls to loadEvent will work even if the supplied file is a json file. Originally, this was precent in analysis.py, but there are other calls to loadEvent, and we should preserve file compatibility for lightcurve file loading between different functions in nmma.