Changed the definition of the prior in gwem-resampling to account for correlation between chirp mass and mass ratio (and potentially spin in NSBH systems).
Added maximum-mass-constraint as a command, so one can sample from a joint GW+EM posterior and get an upper limit on the TOV mass.

The TOV solver here is different from the ones already present in
eos/tov.py, because here we are interested in the density profile of the
star, whereas the other only provides bulk properties.

Conversion factors are now fixed to astropy values.